Rpi-key aggregated cryptosystem/member/requirements.txt
adafruit-blinka
adafruit-circuitpython-atecc
adafruit-circuitpython-bme280
luma.oled
pillow
mfrc522
RPi.GPIO
pycryptodomex
requests
